Tbranch

+++ src/pages/UserList.jsx
@@ -2,6 +2,7 @@ import React from "react";
 import { useEffect, useState } from "react";
 import User from "../components/User";
 import { getUsers } from "../apis/Api";
+import { toast } from "react-toastify";
 
 export default function UserList() {
   const [userlist, setUserlist] = useState([]);
@@ -15,16 +16,23 @@ export default function UserList() {
   }
 
   const handleRemove = async (id) => {
+    console.log("handleremoveclick", handleRemove);
+    const userConfirmed = window.confirm(
+      "Are you sure you want to remove this user?"
+    );
+    if (!userConfirmed) return;
     try {
-      const response = await fetch(`http://localhost:3000/users/${id}`, {
+      const response = await fetch(`http://localhost:3000/users/${id}/delete`, {
         method: "DELETE",
         headers: {
           "Content-Type": "application/json",
         },
       });
+      const data = await response.json();
 
       if (response.ok) {
-        setUserlist(userlist.filter((user) => user.id !== id));
+        toast(data.message);
+        setUserlist(list.filter((user) => user.id !== id));
       } else {
         console.error("Failed to remove user");
       }
@@ -34,9 +42,9 @@ export default function UserList() {
   };
 
   return (
-    <div className="flex flex-col justify-items-center  bg-gray-500 dark:bg-gray-800 dark:border-gray-700 text-gray-900 dark:text-white">
+    <div className="min-h-dvh h-auto flex flex-col justify-items-center bg-gray-500 dark:bg-gray-800 dark:border-gray-700 text-gray-900 dark:text-white">
       <h1 className="m-10 text-center text-xl">USERS</h1>
-      <div className="container mx-auto p-4">
+      <div className="flex justify-self-center mx-auto">
         <div className="grid grid-cols-1 sm:grid-cols-1 md:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4">
           {userlist.map((user, id) => (
             <div key={id} className="p-2">
